basic image viewer tests

+++ assets/shaders/colormaps.glsl
@@ -1,9 +1,9 @@
 // Original copyright info from https://www.shadertoy.com/view/3lBXR3:
 //
-// update of https://www.shadertoy.com/view/WlfXRN to add Google's 
+// update of https://www.shadertoy.com/view/WlfXRN to add Google's
 // turbo colormap
 //
-// License CC0 (public domain) 
+// License CC0 (public domain)
 //   https://creativecommons.org/share-your-work/public-domain/cc0/
 //
 // feel free to use these in your own work!
@@ -17,15 +17,16 @@
 // data fitted from https://github.com/BIDS/colormap/blob/master/colormaps.py
 // (which is licensed CC0)
 //
-// improved version adds turbo colormap from 
+// improved version adds turbo colormap from
 // https://ai.googleblog.com/2019/08/turbo-improved-rainbow-colormap-for.html
 // and https://gist.github.com/mikhailov-work/ee72ba4191942acecc03fe6da94fc73f
 //
 // see also maeln's shader at https://www.shadertoy.com/view/3t2XzV demonstrating turbo
 
-precision highp float;
+precision mediump float;
 
-vec3 viridis(float t) {
+vec3 viridis(float t)
+{
 
     const vec3 c0 = vec3(0.2777273272234177, 0.005407344544966578, 0.3340998053353061);
     const vec3 c1 = vec3(0.1050930431085774, 1.404613529898575, 1.384590162594685);
@@ -35,10 +36,11 @@ vec3 viridis(float t) {
     const vec3 c5 = vec3(4.776384997670288, -13.74514537774601, -65.35303263337234);
     const vec3 c6 = vec3(-5.435455855934631, 4.645852612178535, 26.3124352495832);
 
-    return c0+t*(c1+t*(c2+t*(c3+t*(c4+t*(c5+t*c6)))));
+    return c0 + t * (c1 + t * (c2 + t * (c3 + t * (c4 + t * (c5 + t * c6)))));
 }
 
-vec3 plasma(float t) {
+vec3 plasma(float t)
+{
 
     const vec3 c0 = vec3(0.05873234392399702, 0.02333670892565664, 0.5433401826748754);
     const vec3 c1 = vec3(2.176514634195958, 0.2383834171260182, 0.7539604599784036);
@@ -48,10 +50,11 @@ vec3 plasma(float t) {
     const vec3 c5 = vec3(10.02306557647065, 71.41361770095349, -54.07218655560067);
     const vec3 c6 = vec3(-3.658713842777788, -22.93153465461149, 18.19190778539828);
 
-    return c0+t*(c1+t*(c2+t*(c3+t*(c4+t*(c5+t*c6)))));
+    return c0 + t * (c1 + t * (c2 + t * (c3 + t * (c4 + t * (c5 + t * c6)))));
 }
 
-vec3 magma(float t) {
+vec3 magma(float t)
+{
 
     const vec3 c0 = vec3(-0.002136485053939582, -0.000749655052795221, -0.005386127855323933);
     const vec3 c1 = vec3(0.2516605407371642, 0.6775232436837668, 2.494026599312351);
@@ -61,10 +64,11 @@ vec3 magma(float t) {
     const vec3 c5 = vec3(-50.76852536473588, 29.04658282127291, 4.23415299384598);
     const vec3 c6 = vec3(18.65570506591883, -11.48977351997711, -5.601961508734096);
 
-    return c0+t*(c1+t*(c2+t*(c3+t*(c4+t*(c5+t*c6)))));
+    return c0 + t * (c1 + t * (c2 + t * (c3 + t * (c4 + t * (c5 + t * c6)))));
 }
 
-vec3 inferno(float t) {
+vec3 inferno(float t)
+{
 
     const vec3 c0 = vec3(0.0002189403691192265, 0.001651004631001012, -0.01948089843709184);
     const vec3 c1 = vec3(0.1065134194856116, 0.5639564367884091, 3.932712388889277);
@@ -74,5 +78,5 @@ vec3 inferno(float t) {
     const vec3 c5 = vec3(-71.31942824499214, 32.62606426397723, 73.20951985803202);
     const vec3 c6 = vec3(25.13112622477341, -12.24266895238567, -23.07032500287172);
 
-    return c0+t*(c1+t*(c2+t*(c3+t*(c4+t*(c5+t*c6)))));
+    return c0 + t * (c1 + t * (c2 + t * (c3 + t * (c4 + t * (c5 + t * c6)))));
 }
